refactor(player): decompose VideoPlayer.tsx into focused modules
The last monster file: 1,283 -> 805 lines, zero behavior change.

- useSubtitleManager (274): the entire subtitle cluster — CC toggle
  flow, forced-subtitles auto-fetch, blob lifecycle — verbatim.
  vttContent still flows to SubtitleOverlay, the cast integration
  (subtitleVtt) and the C shortcut.
- useKeyboardShortcuts (110): the single stable document listener
  (Space/K/arrows/M/F/C/Esc) with the same guards.
- PlayerSettingsMenu (171): gear menu (quality badge, movie versions,
  speed) generic over SwitchableContent; visibility state stays in the
  parent so unmount semantics are unchanged.
- ForcedSubtitlesMenu (102): the forced-subs session toggle dropdown.

Bottom controls bar intentionally NOT extracted: it would need 20+
props (controls/state/volume/progress/nav/cast), past the point where
extraction helps.

Prop signature, generics and the memo-with-generic export pattern
preserved exactly.
